SWANDROID-3085: [PM] Cannot wakeup after system resume/suspend from WOL
Part 1 of 2 changes:
===================

Bumped driver version to 5.0 as API has changed to allow a partial
resume power event to be delivered to user-space clients.  This is
necessary to allow the system to partially wake-up, whereby the
kernel and Nexus are brought out of the suspended state, but
Android is still in the sleep/standby state.   By providing the
new "DROID_PM_EVENT_RESUMED_PARTIAL" power event, the libpower HAL
can wake-up Nexus to allow the system to then subsequently suspend
without hangs.

The "full_wol_wakeup" attribute's store function now correctly
returns a valid value.

refs #SWANDROID-3085

Change-Id: I8cf72dbfcd1de8d2285f16899a755493229222d2