droid_pm: Set shutdown flag in pm shutdown notify
This change sets the "shutdown_in_progress" flag in the pm shutdown
notify callback before sending the shutdown request to Nexus. This
flag is an indication to the kernel that an Android shutdown is in
progress which it uses to ensure a clean powerdown.

refs #SWANDROID-4853

Change-Id: Ia692a21f528dca27837863265ac54e44c446a493
Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>
(cherry picked from commit 46f90bf33fd34fbfeb09ca85ee953582211865ff)